+++ projects/swimlane/ngx-ui/src/lib/components/filter/filter.items-position-types.enum.ts
@@ -1,0 +1,4 @@
+export enum FilterItemPositionTypes {
+  Left = 'left',
+  Right = 'right'
+}