+++ src/app/components/filters-page/filters-page.module.ts
@@ -2,20 +2,20 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
-import { FilterSelectModule, SectionModule, SelectModule, TabsModule } from '@swimlane/ngx-ui';
+import { FiltersModule, SectionModule, SelectModule, TabsModule } from '@swimlane/ngx-ui';
 import { PrismModule } from '../../common/prism/prism.module';
 
 import { SelectsPageRoutingModule } from './filters-page-routing.module';
-import { FiltersPageComponent } from './filters-page.component';
+import { FiltersPageComponent, MyCustomComponent } from './filters-page.component';
 
 @NgModule({
-  declarations: [FiltersPageComponent],
+  declarations: [FiltersPageComponent, MyCustomComponent],
   imports: [
     CommonModule,
     FormsModule,
     PrismModule,
     SectionModule,
-    FilterSelectModule,
+    FiltersModule,
     SelectModule,
     SelectsPageRoutingModule,
     FormsModule,