+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## HEAD (unreleased)
 
+## 50.0.0-alpha.0 (2025-06-24)
+
+- Feature: Added support for Angular 20
+- Breaking: Removed support for Angular 17
+- Breaking: Renamed ngx-filter-select to ngx-filter
+- Enhancement (`ngx-filter`): Added support to render an icon and set its position
+- Enhancement (`ngx-filter`): Added support to display the options popup at the right or left of the select
+- Enhancement (`ngx-filter`): Added support to select the SelectAll option automatically when all the options are selected
+- Enhancement (`ngx-filter`): Added support to set a filter type (select, button and customDropdown)
+- Enhancement (`ngx-filter`): Added a new output to emit an event when the filter is clicked
+
 ## 49.3.0 (2025-06-18)
 
 - Feature (`ngx-calendar`): Added new `@Input` named `showStartTimeInputs` and `showEndTimeInputs` set to a `boolean`